feat: move the api call into a github api service
