+++ packages/playground/src/app/simple-page/simple-page.component.ts
@@ -1,16 +1,7 @@
 import { CommonModule } from '@angular/common';
-import { HttpClient } from '@angular/common/http';
-import { Component, inject } from '@angular/core';
-import { QueryProvider } from '@ngneat/ng-query';
+import { Component } from '@angular/core';
 import { SubscribeModule } from '@ngneat/subscribe';
-
-interface Repository {
-  name: string;
-  description: string;
-  subscribers_count: number;
-  stargazers_count: number;
-  forks_count: number;
-}
+import { GithubApiService } from '../github.service';
 
 @Component({
   standalone: true,
@@ -21,21 +12,19 @@ interface Repository {
       <p *ngIf="repo.error">An error has occurred: {{ repo.error }}</p>
 
       <div *ngIf="repo.isSuccess">
-        <h1>{{repo.data.name }}</h1>
-        <p>{{repo.data.description }}</p>
-        <strong>👀 {{repo.data.subscribers_count }}</strong>
-        <strong>✨ {{repo.data.stargazers_count }}</strong>
-        <strong>🍴 {{repo.data.forks_count }}</strong>
-        <div>{{repo.isFetching ? "Updating..." :  ""}}</div>
+        <h1>{{ repo.data.name }}</h1>
+        <p>{{ repo.data.description }}</p>
+        <strong>👀 {{ repo.data.subscribers_count }}</strong>
+        <strong>✨ {{ repo.data.stargazers_count }}</strong>
+        <strong>🍴 {{ repo.data.forks_count }}</strong>
       </div>
+      
+      <div *ngIf="repo.isFetching">Updating...</div>
     </ng-container>
   `,
 })
 export class SimplePageComponent {
-  useQuery = inject(QueryProvider);
-  http = inject(HttpClient);
+  constructor(private githubApiService: GithubApiService) {}
 
-  repo$ = this.useQuery(['repoData'], () =>
-    this.http.get<Repository>('https://api.github.com/repos/ngneat/ng-query')
-  );
+  repo$ = this.githubApiService.getRepository('ngneat/ng-query');
 }
